requirements.txt
annotated-types==0.6.0
anyio==4.0.0
autoflake==1.7.5
black==23.10.1
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.0
flake8==6.1.0
gunicorn==21.2.0
h11==0.14.0
idna==3.4
isort==5.12.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
pip==23.3.1
platformdirs==3.11.0
prometheus-client==0.17.1
prometheus-fastapi-instrumentator==6.1.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
PyYAML==6.0.1
sentry-sdk==1.31.0
setuptools==68.2.2
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.0.5
uvicorn==0.24.0
wheel==0.41.3
